# docs/swagger.yaml
openapi: 3.0.2
info:
  title: barBank
  version: 1.0.0
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `user` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `user` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      tags:
        - users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                'User validation failed: {param}: Path `{param}` is required.':
                  value:
                    error: 'User validation failed: {param}: Path `{param}` is required.'
          description: Input validation failed. Some parameter is missing or in incorrect format.
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Username already exists:
                  value:
                    error: Username already exists
          description: Username already exists
      operationId: createuser
      summary: Create a user
      description: Creates a new instance of a `user`.
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/session'
            examples:
              POST /sessions request body:
                value:
                  username: kalle
                  password: kaalikaskalle
        required: true
      tags:
        - sessions
      responses:
        '201':
          content:
            application/json:
              examples:
                POST /sessions 201 response:
                  value:
                    token: 615bdce137135576c9f489bf
          description: Successful response.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid credentials:
                  value:
                    error: Invalid credentials
          description: Invalid credentials
      operationId: createsession
      summary: Create a session
      description: Creates a new instance of a `session`.
    delete:
      tags:
        - sessions
      responses:
        '204':
          description: Successful logout
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid Authorization header:
                  value:
                    error: Invalid authorization header
                Invalid Authorization header format:
                  value:
                    error: Invalid authorization header format
                Missing Authorization header:
                  value:
                    error: Missing Authorization header
                Invalid token:
                  value:
                    error: Invalid token
          description: Invalid Authorization header
      security:
        -
          Bearer: []
  /users/current:
    get:
      tags:
        - users
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
              examples:
                User details:
                  value:
                    accounts:
                      -
                        balance: 10000
                        createdAt: '2021-09-14T22:38:05.659Z'
                        currency: USD
                        id: 61412463f1f9c72a7962229d
                        name: Main
                        number: abc154ea0a329257241c77623fe2b268c94
                        userId: 61412463f1f9c72a7962229b
                    id: 61412463f1f9c72a7962229b
                    name: Meelis
                    username: poisike
          description: Returns user's details.
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing authorization header:
                  value:
                    error: Missing authorization header
                Invalid authorization header:
                  value:
                    error: Invalid authorization header
          description: Unauthorized
      security:
        -
          Bearer: []
      summary: Get user's details
      description: Returns instance of `user`.
  /transactions:
    summary: Path used to manage the list of transactions.
    description: >-
      The REST endpoint/path used to list and create zero or more `transaction` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `transaction` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/transaction'
        required: true
      tags:
        - transactions
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid amount:
                  value:
                    error: Invalid amount
                Missing parameter:
                  value:
                    error: 'Transaction validation failed:.*:Path`.*` is required.'
                Parameter shorter than the allowed length:
                  value:
                    error: Path `.*` (`.*`) is shorter than the allowed length
          description: Invalid amount
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing authorization header:
                  value:
                    error: Missing authorization header
                Invalid authorization header:
                  value:
                    error: Invalid authorization header
          description: Problem with authorization
        '402':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Insufficient funds:
                  value:
                    error: Insufficient funds
          description: Insufficient funds
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                accountFrom not found:
                  value:
                    error: accountFrom not found
          description: Not found
      security:
        -
          Bearer: []
      operationId: createtransaction
      summary: Create a transaction
      description: Creates a new instance of a `transaction`.
  /transactions/b2b:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/jwt'
        required: true
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/transaction'
              examples:
                POST /transactions/b2b response:
                  value: |-
                    {
                        "receiverName": "Malle Maasikas",
                    }
          description: Transaction accepted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid signature:
                  value:
                    error: Invalid signature
          description: Invalid signature
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Bank \w+ was not found in Central Bank:
                  value: |-
                    {
                        "error": "Bank \w+ was not found in Central Bank"
                    }
          description: Bank was not found in Central Bank"
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                There was an error communicating with the Central Bank:
                  value:
                    error: There was an error communicating with the Central Bank.*
          description: There was an error communicating with Central Bank
      summary: Accepts transactions from other banks
  /transactions/jwks:
    description: ''
    get:
      tags:
        - transactions
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jwks'
          description: >-
            Set of keys in JWKS format containing the public key used to verify JSON Web Token
            sent by the origin bank when initiating a transaction with another bank and signed
            with the corresponding private key
      summary: Returns JWKS for validating jwt signature
      description: >-
        Returns set of keys containing the public key used to verify JSON Web Token sent by the origin
        bank when initiating a transaction with another bank and signed with the corresponding private
        key
components:
  schemas:
    user:
      title: Root Type for user
      description: ''
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
      example:
        username: kalle
        password: kaalikaskalle
        name: Kalle Kaalikas
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: Error message.
    session:
      description: ''
      type: object
    transaction:
      title: Root Type for transaction
      description: ''
      type: object
      properties:
        accountFrom:
          type: string
        accountTo:
          type: string
        amount:
          format: int32
          type: integer
        explanation:
          type: string
      example:
        accountFrom: 132340f6951e3bf584794f409d29f525ef8
        accountTo: 100BY000000001
        amount: 150
        explanation: Don’t go spend it all at once
    jwt:
      title: Root Type for jwt
      description: ''
      type: object
      properties:
        jwt:
          type: string
      example:
        jwt: >-
          eyJhbGciOiJSUzI1NiIsImtpZCI6Ii03VFJQUVdOcFVKdHFSZkplRWdSUzZUOVQxcnl3cHQ4ZE9iam5pM25FeFEifQ.eyJhY2NvdW50RnJvbSI6IjEzMjM0MGY2OTUxZTNiZjU4NDc5NGY0MDlkMjlmNTI1ZWY4IiwiYWNjb3VudFRvIjoiMjIyYWUzZTcyN2ZhZWUyNTA4N2EzMzI3ZTA3MjM4OTQ5ZGQiLCJhbW91bnQiOjIwLCJjdXJyZW5jeSI6IlVTRCIsImV4cGxhbmF0aW9uIjoiYSIsInNlbmRlck5hbWUiOiJLYWxsZSBLYWFsaWthcyJ9.FWT8iV4URKboMjzP_YsSz3c6u84ocZq3xAtbMsJhC1qkFBT0jgr3ZirsUVeN_XKus_26OBKm03KlmDMDtbxagpO_cyJg-ACcW4kRHDS_oJfgbbXAyfoLHBPPpoEisbEJaU5Ijl_8UsTnnJtZtFSLpwG06cLHL7G6edsm01GVRcLmy9Svi9DyIaWXH9tjp2p4n6MmCuugqzmy9DD8f8tl8w-KLH9nAUW5uJK3SOxKByxl5uDEO_TidoMxPWQX9e1aRgGsosACdJbC5R5YVNDw8f6KxZBSzIs6HdkJsBssQ9T7xm3BLCMK8-kqeeeOn7bCJ7WQTvgNSRrqOOOK-OLmHQ
    jwks:
      title: Root Type for jwks
      description: ''
      type: object
      example:
        keys:
          -
            kty: RSA
            kid: QVhlbsNZZY83YRFHS-OWJqiqEMbjoaqhZIiPVRZgW-k
            'n': >-
              RlWISHXpjWXdYGLSB99okpIW8eEL9ochaEMUyZ8bHW46CLXX-kkI8l6ue7KtxuyrWpuOpuzrHtZWp29H6fgyOq1mKZ0heK-XkV1rJoKHd-6wc0Jdpg2SPwC6BWRGcE2HnvHxT0s8pfa_GwTd7Tu27gIR2S7TVx7N86E5jxll7t0
            e: AQAB
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: sessions
    description: ''
  -
    name: users
    description: ''
  -
    name: transactions
    description: ''